Load map and send it to a connected client
Still a lot more to do, this is quite a hack currently.

The map is loaded from the file map.vxl so download a map from aos.party
and save it as map.vxl in the directory where you want to run the server
if you want to try it out.

TODO:
* Handle client disconnecting
* BetterSpades crashes when we send the map